atualizanção do app apa adaptação do usuário adm em andamento

+++ assets/pages/perfil/script.js
@@ -150,12 +150,12 @@ document.addEventListener('DOMContentLoaded', async()=>{
 
     const profile = userRole === 'ADM' ? await getProfileByAdm(userId) : await getProfile()
     
-    /* if(!profile){
+    if(!profile){
         window.alert('Não foi possível carregar os dados do cliente. Efetue login novamente')
         localStorage.clear()
         window.location.href = '../login/index.html'
         return
-    } */
+    }
 
     renderProfile(profile)
     groupedProducts()
